+++ spec/cloud_config_spec.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative '../lib/cloud-config/cache/in_memory'
 require_relative '../lib/cloud-config/providers/in_memory'
 
 RSpec.describe CloudConfig do
@@ -10,6 +11,10 @@ RSpec.describe CloudConfig do
   end
 
   describe '.provider' do
+    it 'initially empty' do
+      expect(config.providers).to be_empty
+    end
+
     context 'without options' do
       before do
         config.configure do
@@ -38,6 +43,32 @@ RSpec.describe CloudConfig do
     end
   end
 
+  describe '.providers_by_key' do
+    it 'initially empty' do
+      expect(config.providers_by_key).to be_empty
+    end
+
+    context 'with settings' do
+      before do
+        config.configure do
+          provider :in_memory do
+            setting :config_key
+          end
+        end
+      end
+
+      it 'contains correct provider for config_key' do
+        expect(config.providers_by_key.keys).to contain_exactly(:config_key)
+        expect(config.providers_by_key[:config_key]).to be_instance_of(CloudConfig::ProviderConfig)
+        expect(config.providers_by_key[:config_key].settings.key?(:config_key)).to be true
+      end
+
+      it 'returns nil for missing key' do
+        expect(config.providers_by_key[:missing]).to be_nil
+      end
+    end
+  end
+
   describe '.cache_client' do
     before do
       config.configure do
@@ -59,12 +90,34 @@ RSpec.describe CloudConfig do
           setting :my_key
         end
       end
+
+      allow(config.providers[:in_memory].provider).to receive(:get).and_return(setting_value)
     end
 
     it 'gets key from provider' do
-      allow(config.providers[:in_memory].provider).to receive(:get).and_return(setting_value)
-
       expect(config.get(:my_key)).to eql(setting_value)
+    end
+
+    context 'with cache' do
+      before do
+        config.configure do
+          cache_client CloudConfig::Cache::InMemory.new
+        end
+      end
+
+      it 'fetches from cache' do
+        expect(config.get(:my_key)).to eql(setting_value)
+        expect(config.get(:my_key)).to eql(setting_value)
+
+        expect(config.providers[:in_memory].provider).to have_received(:get).once
+      end
+
+      it 'resets from cache' do
+        expect(config.get(:my_key)).to eql(setting_value)
+        expect(config.get(:my_key, reset_cache: true)).to eql(setting_value)
+
+        expect(config.providers[:in_memory].provider).to have_received(:get).twice
+      end
     end
   end
 
@@ -81,6 +134,32 @@ RSpec.describe CloudConfig do
       config.set(:my_key, 'key_value')
 
       expect(config.get(:my_key)).to eql('key_value')
+    end
+  end
+
+  describe '.reset' do
+    before do
+      config.configure do
+        cache_client :config_client
+
+        provider :in_memory do
+          setting :my_key
+        end
+      end
+
+      config.reset!
+    end
+
+    it 'sets cache to nil' do
+      expect(config.cache).to be_nil
+    end
+
+    it 'sets providers to be empty' do
+      expect(config.providers).to be_empty
+    end
+
+    it 'sets providers_by_key to be empty' do
+      expect(config.providers_by_key).to be_empty
     end
   end
 end