+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 ## [Unreleased]
 
-## [0.1.0] - 2023-01-15
+## [0.2.0] - 2023-05-20
+
+Bug fixes:
+
+- Fix in memory cache expiry
+
+Enhancements:
+
+- Reset cache for single key
+- Lookup providers by configuration key
+
+## [Released]
+
+## [0.1.0] - 2023-05-17
 
 - Initial release